init token dealer update
