+++ pallets/old-token-dealer/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2018"
+name = "cumulus-token-dealer"
+version = "0.6.0"
+
+[dependencies]
+codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false, features = ["derive"] }
+
+sp-runtime = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "master" }
+frame-support = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "master" }
+frame-system = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "master" }
+
+# Cumulus dependencies
+cumulus-upward-message = { git = "https://github.com/paritytech/cumulus",  default-features = false, branch = "master" }
+cumulus-primitives = { git = "https://github.com/paritytech/cumulus",  default-features = false, branch = "master" }
+
+# Polkadot dependencies
+polkadot-parachain = { git = "https://github.com/paritytech/polkadot", branch = "master", default-features = false }
+
+[features]
+default = ["std"]
+std = [
+    "codec/std",
+    "cumulus-upward-message/std",
+    "cumulus-primitives/std",
+    "sp-runtime/std",
+    "frame-support/std",
+    "frame-system/std",
+    "polkadot-parachain/std",
+]
