+++ pallets/token-dealer/Cargo.toml
@@ -1,31 +1,31 @@
 [package]
-authors = ["Parity Technologies <admin@parity.io>"]
+authors = ["PureStake"]
 edition = "2018"
-name = "cumulus-token-dealer"
-version = "0.6.0"
+name = "token-dealer"
+version = "0.1.0"
 
 [dependencies]
-codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false, features = ["derive"] }
+parity-scale-codec = { version = "2.0.0", default-features = false, features = ["derive"] }
 
-sp-runtime = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "master" }
 frame-support = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "master" }
 frame-system = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "master" }
+sp-runtime = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "master" }
+sp-std = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "master" }
 
-# Cumulus dependencies
-cumulus-upward-message = { git = "https://github.com/paritytech/cumulus",  default-features = false, branch = "master" }
-cumulus-primitives = { git = "https://github.com/paritytech/cumulus",  default-features = false, branch = "master" }
-
-# Polkadot dependencies
-polkadot-parachain = { git = "https://github.com/paritytech/polkadot", branch = "master", default-features = false }
+token-factory = { path = "../token-factory", default-features = false }
+xcm = { git = "https://github.com/paritytech/polkadot", rev = "71a528e822af63b96bdb03efac9b4927fb11a25e", default-features = false }
+xcm-executor = { git = "https://github.com/paritytech/polkadot", rev = "71a528e822af63b96bdb03efac9b4927fb11a25e", default-features = false }
+#cumulus-upward-message = { git = "https://github.com/paritytech/cumulus",  default-features = false, branch = "master" }
+cumulus-primitives = { git = "https://github.com/paritytech/cumulus",  default-features = false, rev = "6438d328b8ac09f43d19df65a2cf8e31a524d78f" }
 
 [features]
 default = ["std"]
 std = [
-    "codec/std",
-    "cumulus-upward-message/std",
-    "cumulus-primitives/std",
-    "sp-runtime/std",
+    "parity-scale-codec/std",
     "frame-support/std",
     "frame-system/std",
-    "polkadot-parachain/std",
+    "sp-runtime/std",
+    "sp-std/std",
+    "token-factory/std",
+    "cumulus-primitives/std",
 ]
