--- pallets/x-transfer/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-authors = ["PureStake"]
-edition = "2018"
-name = "x-transfer"
-version = "0.1.0"
-
-[dependencies]
-parity-scale-codec = { version = "2.0.0", default-features = false, features = ["derive"] }
-
-frame-support = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "master" }
-frame-system = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "master" }
-sp-runtime = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "master" }
-sp-std = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "master" }
-
-token-factory = { path = "../token-factory", default-features = false }
-
-[features]
-default = ["std"]
-std = [
-    "parity-scale-codec/std",
-    "frame-support/std",
-    "frame-system/std",
-    "sp-runtime/std",
-    "sp-std/std",
-    "token-factory/std",
-]
