--- pallets/x-transfer/src/lib.rs
@@ -1,58 +1,0 @@
-// Copyright 2019-2020 PureStake Inc.
-// This file is part of Moonbeam.
-
-// Moonbeam is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// Moonbeam is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with Moonbeam.  If not, see <http://www.gnu.org/licenses/>.
-
-//! Minimal Pallet that stores the numeric Ethereum-style chain id in the runtime.
-
-#![cfg_attr(not(feature = "std"), no_std)]
-
-use frame_support::pallet;
-
-pub use pallet::*;
-
-#[pallet]
-pub mod pallet {
-	use erc_factory::TokenFactory;
-	use frame_support::{pallet_prelude::*, traits::Currency};
-	use frame_system::pallet_prelude::*;
-
-	pub enum Currencies {
-		Native,
-		ERC20(TokenId),
-	}
-
-	type TokenId = Vec<u8>;
-	type BalanceOf<T> =
-		<<T as Config>::Currency as Currency<<T as frame_system::Config>::AccountId>>::Balance;
-
-	/// Pallet for executing cross-chain transfers
-	#[pallet::pallet]
-	pub struct Pallet<T>(PhantomData<T>);
-
-	/// Configuration trait of this pallet.
-	#[pallet::config]
-	pub trait Config: frame_system::Config {
-		/// Native balances type
-		type Currency: Currency<Self::AccountId>;
-		/// Abstraction over EVM to register, mint, and burn ERC20 tokens
-		type ErcFactory: TokenFactory<TokenId, Self::AccountId, BalanceOf<Self>>;
-	}
-
-	#[pallet::hooks]
-	impl<T: Config> Hooks<BlockNumberFor<T>> for Pallet<T> {}
-
-	#[pallet::call]
-	impl<T: Config> Pallet<T> {}
-}
